// TSIC-Core-Angular/src/frontend/tsic-app/src/app/core/api/models/FeeModifierDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FeeModifierDto = {
    feeModifierId?: string | null;
    modifierType: string;
    amount: number;
    startDate?: string | null;
    endDate?: string | null;
};